Support adding descriptions to controls.
Controls are now able to be given a custom description to provide better
context when viewing Storybook.js documentation.

Adding a description works in the same way as adding a custom name for
the control — it can be set when initializing the control or by calling
the `description` method on the instance.